feat: make deepsearch-toolkit optional
